requirements.txt
python-registry
tabulate